add multi project support and in-engine project source build support
There is now a dropdown to select projects if we've found several. This will also read out projects inside engine source builds, to support that structure as well. 
It will now also not open the plugin manager window if the solution seems to be not one of an unreal engine project of any kind